+++ src/data/configs.js
@@ -45,6 +45,7 @@ var configs = {
     'delayToShowTitleOnHoverWhenHidden': 600, // ms
     'circleShadowOpacity': 0.3,
     'circleLocation': 'cursorOverflow', /// possible values: 'alwaysCursor', 'alwaysCorner', cursorCorner', 'cursorOverflow
+    'longLeftClickThreshold': 21,
 
     /// Menus
     'regularMenu': {